+++ ReleaseNotes.md
@@ -19,3 +19,4 @@
 - Add MILE_TOKEN_PRIVILEGE_LUID_CLASS enum.
 - Add MileAllocateMemory function.
 - Add MileReallocateMemory function.
+- Add MileFreeMemory function.